geocode before adding step
